requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
aiosqlite==0.20.0
python-multipart==0.0.12
anthropic==0.40.0
pdfplumber==0.11.4
openpyxl==3.1.5
pandas==2.2.3
python-dotenv==1.0.1
httpx==0.27.2
Pillow==10.4.0
